requirements.txt
djangorestframework==3.14.0
python-dotenv==1.0.0
drf_yasg==1.21.6
psycopg2-binary==2.9.9
